Ensure that parsed values from location are stable
