+++ app/scripts/components/exploration/atoms/types.d.ts
@@ -1,0 +1,1 @@
+export type AtomValueUpdater<T> = T | ((prev: T) => T);
